Updated schema to accept settings and env_vars
