+++ scale/job/execution/running/tasks/base_task.py
@@ -98,7 +98,7 @@ class Task(object):
         """Returns the Docker parameters used to run this task
 
         :returns: The Docker parameters
-        :rtype: [:class:`job.configuration.configuration.job_configuration.DockerParam`]
+        :rtype: [:class:`job.configuration.configuration.job_parameter.DockerParam`]
         """
 
         return self._docker_params
